Fixed go 1.9 tests
